requirements.txt
fastapi
uvicorn
pyyaml
jinja2
python-multipart
